+++ CHANGELOG.md
@@ -35,6 +35,7 @@
 - Updated Yii to 2.0.40.
 
 ### Deprecated
+- Deprecated `craft\base\VolumeInterface::deleteDir()`. `deleteDirectory()` should be used instead.
 - Deprecated `craft\base\VolumeInterface::getFileMetadata()`. `getFileSize()` and `getDateModified()` should be used instead.
 - Deprecated `craft\helpers\Template::paginateCriteria()`. `paginateQuery()` should be used instead.